change(web): converts most of kmwembedded.ts's remaining entries
